move function to utils object
